requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv